backend/requirements.txt
fastapi==0.115.4
httpx==0.27.2
python-dotenv==1.0.1
pyodbc==5.2.0
uvicorn==0.32.0